iot-simulation/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
numpy==1.24.3
pandas==2.0.3
faker==19.3.1
geopy==2.3.0
colorama==0.4.6
click==8.1.6
pydantic==2.1.1
asyncio==3.4.3
aiohttp==3.8.5
